定义task接口，实现队列。

+++ src/main/java/com/github/skyisbule/db/task/IoTask.java
@@ -5,7 +5,7 @@ import com.github.skyisbule.db.type.IoTaskType;
 //定义io任务的结构
 public class IoTask {
 
-    public long       transactionId;
+    public int       transactionId;
     public String     file;
     public IoTaskType type;
     public Integer    offset;
